rename: 후보지 추가 관련 컴포넌트 폴더 이동
